數字加入千分位

+++ flasky/app/models/trade_records.py
@@ -10,8 +10,8 @@ class TradeRecords(DBAbstract):
     user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
     stock_basic_info_id = db.Column(db.Integer, db.ForeignKey('stock_basic_info.id'), nullable=False)
     volume = db.Column(db.Integer, nullable=False)
-    price = db.Column(db.Integer, nullable=False)
-    cost = db.Column(db.Integer, nullable=False)
+    price = db.Column(db.Float, nullable=False)
+    cost = db.Column(db.Float, nullable=False)
     created_at = db.Column(db.DateTime(), default=datetime.now)
     updated_at = db.Column(db.DateTime(), default=datetime.now)
